v2.0.2

+++ src/app/auth/services/auth.service.ts
@@ -1,12 +1,23 @@
 import { Injectable } from '@angular/core';
+import { GoogleAuthProvider } from '@angular/fire/auth';
 import { AngularFireAuth } from '@angular/fire/compat/auth';
 import { AngularFirestore, AngularFirestoreDocument } from '@angular/fire/compat/firestore';
 import { Router } from '@angular/router';
-import { SnackbarService } from '@lib/services/snackbar/snackbar.service';
-import firebase from 'firebase/compat';
-import { BehaviorSubject } from 'rxjs';
+import { LoggerFactory } from '@lib/helpers/logger.factory';
+import { asUniqueArray } from '@lib/helpers/unique-array';
+import { ERole, TRole } from '@lib/models/role';
+import firebase from 'firebase/compat/app';
+import { BehaviorSubject, from, Observable, switchMap, take } from 'rxjs';
 
-export type FirebaseAuth = firebase.auth.UserCredential;
+
+export interface AuthUser {
+  uid: string;
+  email: string;
+  expire: number;
+  roles: TRole[];
+}
+
+type FirebaseAuth = firebase.auth.UserCredential;
 
 @Injectable({
   providedIn: 'root'
@@ -14,17 +25,14 @@ export type FirebaseAuth = firebase.auth.UserCredential;
 export class AuthService {
 
   isLoggedIn$ = new BehaviorSubject<boolean>(false);
+  private logger = this.loggerFactory.createLogger('Authentication', 'service');
 
-  private userSubj$ = new BehaviorSubject<FirebaseAuth>(null);
+  private userSubj$ = new BehaviorSubject<AuthUser>(null);
   readonly user$ = this.userSubj$.asObservable();
 
-  private readonly TOKEN_KEY = 'firebase_auth';
+  private readonly TOKEN_KEY = 'auth';
 
-  get token(): string {
-    return localStorage.getItem(this.TOKEN_KEY);
-  }
-
-  get user(): FirebaseAuth {
+  get user(): AuthUser {
     return this.userSubj$.value;
   }
 
@@ -32,36 +40,57 @@ export class AuthService {
     private readonly router: Router,
     private readonly afa: AngularFireAuth,
     private readonly afs: AngularFirestore,
-    private readonly snackbar: SnackbarService,
+    private readonly loggerFactory: LoggerFactory
   ) {
-    this.isLoggedIn$.next(!!this.token);
+    this.handleState();
   }
 
-  login({ email, password }): Promise<void> {
-    return this.afa.signInWithEmailAndPassword(email, password)
-      .then((response: FirebaseAuth) => {
-        this.setUserData(response.user);
-        this.isLoggedIn$.next(true);
-        this.snackbar.success('Logged in successfully.');
-      })
-      .catch((error) => this.snackbar.error(error));
+  private handleState(): void {
+    // TODO: check the user is existed in collection
+    const current = JSON.parse(localStorage.getItem(this.TOKEN_KEY));
+
+    if (current) {
+      const expired = Date.now() >= current.expire * 1000;
+      return this.updateUserState(current, !expired);
+    };
+    return this.updateUserState(null, false);
   }
 
-  register({ email, password }): Promise<void> {
-    return this.afa.createUserWithEmailAndPassword(email, password)
-      .then((response: FirebaseAuth) => {
-        this.setUserData(response.user);
-        this.snackbar.success('Logged in successfully.');
-        this.verifyEmail();
-      })
-      .catch((error) => this.snackbar.error(error));
+  /* Sign in */
+
+  emailSignIn({ email, password }): Observable<void> {
+    return from(this.afa.signInWithEmailAndPassword(email, password)).pipe(
+      take(1),
+      switchMap((response: FirebaseAuth) => this.updateUserData(response.user))
+    );
   }
 
-  resetPassword(password: string): void {
-    this.afa.sendPasswordResetEmail(password)
-      .then(() => this.snackbar.success('Password reset was sent, please check your message box.'))
-      .catch((error) => this.snackbar.error(error));
+  googleSignIn(): Observable<void> {
+    return from(this.afa.signInWithPopup(new GoogleAuthProvider())).pipe(
+      take(1),
+      switchMap((response: FirebaseAuth) => this.updateUserData(response.user))
+    );
   }
+
+  /* Register */
+
+  emailRegister({ email, password }): Observable<void> {
+    return from(this.afa.createUserWithEmailAndPassword(email, password)).pipe(
+      switchMap((response: FirebaseAuth) => this.updateUserData(response.user))
+    );
+  }
+
+  resetPassword(password: string): Observable<void> {
+    return from(this.afa.sendPasswordResetEmail(password));
+  }
+
+  async logout(): Promise<void> {
+    await this.afa.signOut();
+    localStorage.clear();
+    this.updateUserState(null, false);
+  }
+
+  /* Utils */
 
   private verifyEmail(): Promise<boolean> {
     return this.afa.currentUser
@@ -69,16 +98,37 @@ export class AuthService {
       .then(() => this.router.navigate(['verify-email-address']));
   }
 
-  setUserData(user: FirebaseAuth['user']): Promise<void> {
-    const ref: AngularFirestoreDocument = this.afs.doc(
-      `users/${user.uid}`
-    );
-    localStorage.setItem(this.TOKEN_KEY, JSON.stringify(user));
-    return ref.set(user, { merge: true });
+  private async updateUserData(user: firebase.User): Promise<void> {
+    if (!user) {
+      throw new Error('No user found');
+    }
+
+    const { uid, email } = user;
+    const idToken = await user.getIdToken();
+    const expire = JSON.parse(
+      Buffer
+        .from(idToken.split('.')[1], 'base64')
+        .toString('ascii')
+    )['exp'];
+
+    const userRef: AngularFirestoreDocument<AuthUser> = this.afs.doc(`users/${uid}`);
+    const data: AuthUser = {
+      uid,
+      email,
+      expire,
+      roles: asUniqueArray([ERole.SUBSCRIBER])
+    };
+    this.updateUserState(data, true);
+    localStorage.setItem(this.TOKEN_KEY, JSON.stringify(data));
+    return userRef.set(data, { merge: true });
   }
 
-  logout(): void {
-    this.isLoggedIn$.next(false);
-    localStorage.clear();
+  updateUserState(user: AuthUser, loggedIn: boolean): void {
+    if (!loggedIn) {
+      localStorage.clear();
+    }
+    this.userSubj$.next(user);
+    this.isLoggedIn$.next(loggedIn);
+    this.logger.log(user ? new Date(user.expire * 1000) : 'Not signed in yet');
   }
 }
